e2e cleaning up
