+++ packages/nx-firebase/src/generators/sync/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './tags'
