+++ packages/nx-firebase/src/generators/sync/sync.ts
@@ -1,5 +1,4 @@
 import {
-  formatFiles,
   GeneratorCallback,
   getProjects,
   logger,
@@ -7,9 +6,7 @@ import {
   readJson,
   readProjectConfiguration,
   runTasksInSerial,
-  TargetConfiguration,
   Tree,
-  updateJson,
   updateProjectConfiguration,
   writeJson,
 } from '@nx/devkit'
@@ -19,99 +16,15 @@ import {
   calculateFirebaseConfigName,
   FirebaseConfig,
   FirebaseFunction,
-  generateFirebaseConfigName,
 } from '../../utils'
 import initGenerator from '../init/init'
-
-function debugInfo(info: string) {
-  //  logger.info(info)
-}
-
-function updateFirebaseAppDeployProject(
-  target: TargetConfiguration,
-  options: SyncGeneratorSchema,
-) {
-  const command: string = target.options.command
-  if (command.includes('firebase deploy')) {
-    debugInfo('- found deploy command')
-    debugInfo(`- command=${command}`)
-    if (command.includes('--project')) {
-      debugInfo('- found --project in deploy command')
-      // already set, so update
-      target.options.command = command.replace(
-        /(--project[ =])([A-Z0-9a-z-_]+)/,
-        '$1' + options.project,
-      )
-      debugInfo(`- new command: ${target.options.command}`)
-
-      logger.info(
-        `  SYNC updating firebase deploy project for '${options.app}' to '--project=${options.project}'`,
-      )
-      return true
-    } else {
-      debugInfo('- did not find --project in deploy command')
-
-      // no set, so add
-      target.options.command = command.replace(
-        /(firebase deploy)/,
-        '$1' + ' ' + `--project=${options.project}`,
-      )
-      debugInfo(`- new command: ${target.options.command}`)
-      logger.info(
-        `  SYNC setting firebase deploy project for '${options.app}' to '--project=${options.project}'`,
-      )
-      return true
-    }
-  }
-  return false
-}
-
-// for specific or all apps
-// if firebase app is renamed - all dependent functions need to be re-tagged.
-
-// for each specific or all apps
-// if firebase app is deleted - all functions are orphaned
-//
-//  find all firebase function apps that are matched to app
-//   run through implicit deps
-//
-
-// if firebase function is renamed - app deps & firebase.json need to be updated
-// if firebase function is deleted - app deps & firebase.json need to be updated
-
-// should sync be workspace or per app or per function?
-// sync (all apps & workspaces)
-// sync --functions (just workspace functions)
-// sync --app=<app> [--functions] [--project=xxx] (specific app)
-
-// dont need the functions flags, we'll always sync those
-
-function getFirebaseScopeFromTag(project: ProjectConfiguration, scope: string) {
-  const tags = project.tags
-  let foundTag = false
-  if (tags) {
-    for (let i = 0; i < tags.length; ++i) {
-      const tag = tags[i]
-      if (tag.includes(scope)) {
-        foundTag = true
-        const scopes = tag.split(':')
-        if (scopes.length === 3) {
-          return { tagValue: scopes[2], tagIndex: i }
-        } else {
-          logger.error(
-            `Malformed '${scope}' tag in project '${project.name}', expected '${scope}:<value>', found '${tag}'.`,
-          )
-        }
-      }
-    }
-  }
-  if (!foundTag) {
-    logger.error(
-      `Project '${project.name}' has a missing '${scope}' tag in project.`,
-    )
-  }
-  return undefined
-}
+import { getFirebaseScopeFromTag } from './lib'
+import { debugInfo } from './lib/debug'
+import {
+  getAllFirebaseProjects,
+  isFirebaseApp,
+  updateFirebaseAppDeployProject,
+} from './lib/firebase-projects'
 
 export async function syncGenerator(
   tree: Tree,
@@ -125,66 +38,25 @@ export async function syncGenerator(
 
   // change the firebase project for an nx firebase app project
   if (options.project) {
-    debugInfo(`changing project for ${options.app} to ${options.project}`)
+    // --project option requires --app option to be specified
+    if (!options.app) {
+      throw new Error('--app not specified for --project option')
+    }
+    // validate app parameter
     const project = readProjectConfiguration(tree, options.app)
-    if (!project.tags || project.tags.includes['firebase:app']) {
+    if (!isFirebaseApp(project)) {
       throw new Error(`Project '${options.app}' is not a Firebase application.`)
     }
+    debugInfo(`changing project for ${options.app} to ${options.project}`)
     if (updateFirebaseAppDeployProject(project.targets.deploy, options)) {
       updateProjectConfiguration(tree, options.app, project)
     }
     return
   }
 
-  // otherwise sync either all or one specific firebase app project
-
+  // otherwise, sync the workspace.
   // build list of firebase apps and functions in the workspace
-  const projects = getProjects(tree)
-  let firebaseAppProjects: Record<string, ProjectConfiguration> = {}
-  const firebaseFunctionProjects: Record<string, ProjectConfiguration> = {}
-
-  debugInfo('- building list of firebase apps & functions')
-
-  //   debugInfo(JSON.stringify(tree))
-  //   debugInfo(`projects=${JSON.stringify(projects)}`)
-
-  for (const p of projects) {
-    const projectName = p[0]
-    const project = p[1] //projects[projectName]
-    const tags = project.tags
-
-    debugInfo(`- checking project ${projectName} tags=${JSON.stringify(tags)}`)
-
-    if (tags.includes('firebase:app')) {
-      debugInfo(`- found firebase app ${projectName}`)
-      firebaseAppProjects[projectName] = project
-    }
-    if (tags.includes('firebase:function')) {
-      debugInfo(`- found firebase function ${projectName}`)
-      firebaseFunctionProjects[projectName] = project
-    }
-  }
-
-  debugInfo(
-    `- firebaseAppProjects=${JSON.stringify(Object.keys(firebaseAppProjects))}`,
-  )
-  debugInfo(
-    `- firebaseFunctionProjects=${JSON.stringify(
-      Object.keys(firebaseFunctionProjects),
-    )}`,
-  )
-
-  // if a single application has been specified for sync,
-  // filter the list of firebase app projects to be processed
-  if (options.app) {
-    if (options.app in firebaseAppProjects) {
-      firebaseAppProjects = {
-        [options.app]: firebaseAppProjects[options.app],
-      }
-    } else {
-      throw new Error(`Project '${options.app}' is not a Firebase application.`)
-    }
-  }
+  const workspace = getAllFirebaseProjects(tree)
 
   // build lists of firebase apps & functions that have been deleted or renamed
   debugInfo('- Syncing workspace')
@@ -198,10 +70,11 @@ export async function syncGenerator(
 
   // determine deleted functions
   debugInfo(`- checking apps for deleted function deps`)
-  for (const firebaseAppProjectName in firebaseAppProjects) {
-    const firebaseAppProject = firebaseAppProjects[firebaseAppProjectName]
+  for (const firebaseAppProjectName in workspace.firebaseAppProjects) {
+    const firebaseAppProject =
+      workspace.firebaseAppProjects[firebaseAppProjectName]
     firebaseAppProject.implicitDependencies?.map((dep) => {
-      if (!(dep in firebaseFunctionProjects)) {
+      if (!(dep in workspace.firebaseFunctionProjects)) {
         debugInfo(
           `- function ${dep} is a dep, but cannot be located so deleted`,
         )
@@ -228,7 +101,7 @@ export async function syncGenerator(
     for (let i = 0; i < functions.length; ++i) {
       const func = functions[i]
       const funcName = func.codebase
-      if (!(funcName in firebaseFunctionProjects)) {
+      if (!(funcName in workspace.firebaseFunctionProjects)) {
         debugInfo(
           `- function ${funcName} is in config ${firebaseConfigName}, but cannot be located so function must be deleted`,
         )
@@ -246,14 +119,14 @@ export async function syncGenerator(
   // This is ok though because we detect renamed functions via tags
 
   // determine deleted apps
-  for (const firebaseFunctionName in firebaseFunctionProjects) {
+  for (const firebaseFunctionName in workspace.firebaseFunctionProjects) {
     const firebaseFunctionProject =
-      firebaseFunctionProjects[firebaseFunctionName]
+      workspace.firebaseFunctionProjects[firebaseFunctionName]
     const { tagValue } = getFirebaseScopeFromTag(
       firebaseFunctionProject,
       'firebase:app',
     )
-    if (!(tagValue in firebaseAppProjects)) {
+    if (!(tagValue in workspace.firebaseAppProjects)) {
       debugInfo(
         `- function ${firebaseFunctionName} points to app ${tagValue} which doesnt exist, so app must be deleted`,
       )
@@ -265,8 +138,9 @@ export async function syncGenerator(
   debugInfo(`checking for renamed apps`)
 
   // determine renamed apps
-  for (const firebaseAppProjectName in firebaseAppProjects) {
-    const firebaseAppProject = firebaseAppProjects[firebaseAppProjectName]
+  for (const firebaseAppProjectName in workspace.firebaseAppProjects) {
+    const firebaseAppProject =
+      workspace.firebaseAppProjects[firebaseAppProjectName]
     const { tagValue } = getFirebaseScopeFromTag(
       firebaseAppProject,
       'firebase:name',
@@ -285,9 +159,9 @@ export async function syncGenerator(
   debugInfo(`- checking for renamed functions`)
 
   // determine renamed functions
-  for (const firebaseFunctionName in firebaseFunctionProjects) {
+  for (const firebaseFunctionName in workspace.firebaseFunctionProjects) {
     const firebaseFunctionProject =
-      firebaseFunctionProjects[firebaseFunctionName]
+      workspace.firebaseFunctionProjects[firebaseFunctionName]
     const { tagValue } = getFirebaseScopeFromTag(
       firebaseFunctionProject,
       'firebase:name',
@@ -313,8 +187,8 @@ export async function syncGenerator(
   )
 
   // now sync the selected firebase apps
-  for (const firebaseAppName in firebaseAppProjects) {
-    const firebaseAppProject = firebaseAppProjects[firebaseAppName]
+  for (const firebaseAppName in workspace.firebaseAppProjects) {
+    const firebaseAppProject = workspace.firebaseAppProjects[firebaseAppName]
     // const implicitDependencies: string[] = []
 
     // what if function renamed AND app is renamed?
@@ -371,7 +245,7 @@ export async function syncGenerator(
     firebaseAppProject.implicitDependencies?.map((firebaseFunctionName) => {
       let firebaseFunctionUpdated = false
       const firebaseFunctionProject =
-        firebaseFunctionProjects[firebaseFunctionName]
+        workspace.firebaseFunctionProjects[firebaseFunctionName]
       const { tagValue, tagIndex } = getFirebaseScopeFromTag(
         firebaseFunctionProject,
         'firebase:app',
@@ -461,7 +335,7 @@ export async function syncGenerator(
         const newFuncName = renamedFunctions[funcName]
         func.codebase = newFuncName
         // change source dir
-        const project = firebaseFunctionProjects[newFuncName]
+        const project = workspace.firebaseFunctionProjects[newFuncName]
         func.source = project.targets.build.options.outputPath
         logger.info(
           `  SYNC renamed firebase function codebase from '${funcName}' to '${newFuncName}' in '${firebaseConfigName}'`,
@@ -477,9 +351,9 @@ export async function syncGenerator(
 
   // last, detect and warn about orphaned functions (where app has been deleted)
   debugInfo(`- checking for orphaned functions`)
-  for (const firebaseFunctionName in firebaseFunctionProjects) {
+  for (const firebaseFunctionName in workspace.firebaseFunctionProjects) {
     const firebaseFunctionProject =
-      firebaseFunctionProjects[firebaseFunctionName]
+      workspace.firebaseFunctionProjects[firebaseFunctionName]
     const { tagValue } = getFirebaseScopeFromTag(
       firebaseFunctionProject,
       'firebase:app',
